+++ server/game/src/data/v14/packets/client/admin.rs
@@ -22,6 +22,7 @@ pub struct AdminSendNoticePacket {
     pub level_id: LevelId,
     pub player: FastString,
     pub message: FastString,
+    pub can_reply: bool,
 }
 
 #[derive(Packet, Decodable)]